#94 Refactored PhotosFlickr classes to conform with Architect Coders proposed architecture
