+++ app/src/main/java/com/example/flickrgallery/data/source/PhotosFlickerDb.kt
@@ -1,0 +1,23 @@
+package com.example.flickrgallery.data.source
+
+import com.example.flickrgallery.client.FlickrDbService
+import okhttp3.OkHttpClient
+import okhttp3.logging.HttpLoggingInterceptor
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+
+class PhotosFlickerDb (baseUrl: String) {
+
+    val okHttpClient = HttpLoggingInterceptor().run {
+        level = HttpLoggingInterceptor.Level.BODY
+        OkHttpClient.Builder().addInterceptor(this).build()
+    }
+
+    val service = Retrofit.Builder()
+        .baseUrl(baseUrl)
+        .client(okHttpClient)
+        .addConverterFactory(GsonConverterFactory.create())
+        .build()
+        .run { create(FlickrDbService::class.java) }
+
+}
